+++ auth.go
@@ -1,7 +1,43 @@
-// @Author: abbeymart | Abi Akindele | @Created: 2020-12-04 | @Updated: 2020-12-04
+// @Author: abbeymart | Abi Akindele | @Created: 2020-12-08 | @Updated: 2020-12-08
 // @Company: mConnect.biz | @License: MIT
-// @Description: crud authorization
+// @Description: go: mConnect
 
 package mccrud
 
+import (
+	"fmt"
+	"github.com/abbeymart/mcresponsego"
+)
 
+func (crud Crud) TaskPermission() mcresponse.ResponseMessage {
+	fmt.Println(crud)
+
+	return mcresponse.GetResMessage("paramsError", mcresponse.ResponseMessageOptions{
+		Message: "action-params is required to perform save operation.",
+		Value:   nil,
+	})
+}
+
+func (crud Crud) CheckTaskAccess() mcresponse.ResponseMessage {
+	fmt.Println(crud)
+
+	return mcresponse.GetResMessage("paramsError", mcresponse.ResponseMessageOptions{
+		Message: "action-params is required to perform save operation.",
+		Value:   nil,
+	})
+}
+
+func (crud Crud) GetRoleServices() []RoleServiceType {
+	fmt.Println(crud)
+
+	return []RoleServiceType{}
+}
+
+func (crud Crud) GetCurrentRecord() mcresponse.ResponseMessage {
+	fmt.Println(crud)
+
+	return mcresponse.GetResMessage("paramsError", mcresponse.ResponseMessageOptions{
+		Message: "action-params is required to perform save operation.",
+		Value:   nil,
+	})
+}