+++ getStream.go
@@ -1,20 +1,12 @@
 // @Author: abbeymart | Abi Akindele | @Created: 2020-12-01 | @Updated: 2020-12-01
 // @Company: mConnect.biz | @License: MIT
-// @Description: delete or remove record(s)
+// @Description: get / query - stream record(s)
 
 package mccrud
 
 import "github.com/abbeymart/mcresponsego"
 
-func (crud Crud) Delete() mcresponse.ResponseMessage {
-
-	return mcresponse.GetResMessage("success", mcresponse.ResponseMessageOptions{
-		Message: "success",
-		Value:   nil,
-	})
-}
-
-func (crud Crud) Remove() mcresponse.ResponseMessage {
+func (crud Crud) GetStream() mcresponse.ResponseMessage {
 
 	return mcresponse.GetResMessage("success", mcresponse.ResponseMessageOptions{
 		Message: "success",