+++ helper/computeTempTable.go
@@ -1,0 +1,5 @@
+// @Author: abbeymart | Abi Akindele | @Created: 2020-12-08 | @Updated: 2020-12-08
+// @Company: mConnect.biz | @License: MIT
+// @Description: compute temporary-table script
+
+package helper