+++ save.go
@@ -3,3 +3,37 @@
 // @Description: save (create / update) record(s)
 
 package mccrud
+
+import "github.com/abbeymart/mcresponsego"
+
+func (crud Crud) Save() mcresponse.ResponseMessage {
+
+	return mcresponse.GetResMessage("success", mcresponse.ResponseMessageOptions{
+		Message: "success",
+		Value:   nil,
+	})
+}
+
+func (crud Crud) Create() mcresponse.ResponseMessage {
+
+	return mcresponse.GetResMessage("success", mcresponse.ResponseMessageOptions{
+		Message: "success",
+		Value:   nil,
+	})
+}
+
+func (crud Crud) Update() mcresponse.ResponseMessage {
+
+	return mcresponse.GetResMessage("success", mcresponse.ResponseMessageOptions{
+		Message: "success",
+		Value:   nil,
+	})
+}
+
+func (crud Crud) ComputeRecords() mcresponse.ResponseMessage {
+
+	return mcresponse.GetResMessage("success", mcresponse.ResponseMessageOptions{
+		Message: "success",
+		Value:   nil,
+	})
+}