+++ HISTORY.md
@@ -4,3 +4,4 @@
 
 - Release date: Dec 12, 2020
 - CRUD operations package for RDBMS in Go - PostgreSQL and SQLite
+- Consideration/Optional: add mongoDB package features, as required